update: styles in Cart page

+++ components/FoodCard.jsx
@@ -2,12 +2,18 @@ import React from "react";
 import Image from "next/image";
 
 function FoodCard({ foods }) {
+  
   const handleClick = (order) => {
     const orders = JSON.parse(localStorage.getItem("orders")) || [];
-    const updateOrders = [...orders, order]
-    console.log('in button', orders)
+    const duplicateOrder = orders.find((element) => {
+      return element._id === order._id;
+    });
+
+    if (duplicateOrder) return alert("The product is already in the Cart");
+    const updateOrders = [...orders, order];
     localStorage.setItem("orders", JSON.stringify(updateOrders));
-  }
+  };
+
   return (
     <ul className="grid  grid-cols-1 gap-4 px-2 md:grid-cols-3 md:px-8">
       {foods.map((food) => (
@@ -25,7 +31,10 @@ function FoodCard({ foods }) {
             <p>{food.price} hrn</p>
           </div>
           <div className="flex justify-end">
-            <button className="black_btn" onClick={() => handleClick(food)}>
+            <button
+              className="black_btn"
+              onClick={() => handleClick({ ...food, quantity: 1 })}
+            >
               Add to card
             </button>
           </div>
